+++ app/Plugin/Transients.php
@@ -221,7 +221,6 @@ class Transients {
 	 *
 	 * @return void
 	 * @noinspection PhpUnused
-	 * @noinspection PhpNoReturnAttributeCanBeAddedInspection
 	 */
 	public function dismiss_transient_via_ajax(): void {
 		// check nonce.
@@ -245,7 +244,7 @@ class Transients {
 		self::get_instance()->get_transient_by_name( $option_name )->delete();
 
 		// return nothing.
-		wp_die();
+		wp_send_json_success();
 	}
 
 	/**