+++ changelog.md
@@ -10,10 +10,12 @@
 
 - Changed widget titles
 - Set compatibility to WordPress 6.8
+- Small code optimizations
 
 ### Fixed
 
 - Fixed wrong review URL
+- Fixed missing translation if no widgets are available
 
 ## [1.0.0] - 17.04.2025